requirements.txt
Flask
gunicorn
pymongo
flask-cors
python-dotenv
python-dateutil
dnspython
flasgger